More floats changed to doubles, and added explicit casts to eliminate warnings
git-svn-id: http://svn.code.sf.net/p/petavision/code/trunk@6138 7c0444ca-5550-0410-9008-faecdbe78eb5

+++ src/columns/HyPerCol.hpp
@@ -128,9 +128,9 @@ private:
 
    long int numSteps;
    long int currentStep;
-   int layerArraySize;
+   size_t layerArraySize;
    int numLayers;
-   int connectionArraySize;
+   size_t connectionArraySize;
    int numConnections;
 
    bool warmStart;
